% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapR.R
\name{bootstrapR}
\alias{bootstrapR}
\title{Calculates the bootstrap confidence intervals for the effective reproductive number}
\usage{
bootstrapR(probs, dateVar, indIDVar, pVar, timeFrame = c("days",
  "months", "weeks", "years"), rangeForAvg = NULL, B = 1000,
  alpha = 0.05)
}
\arguments{
\item{probs}{The name of the dateset with transmission probabilities}

\item{dateVar}{The variable name (in quotes) of the dates that the individuals are observed.}

\item{indIDVar}{The variable name (in quotes) of the individual ID.}

\item{pVar}{The variable name (in quotes) of the transmission probabilities.}

\item{timeFrame}{The time frame used to calculate Rt. One of "days", "weeks", "months", "years".}

\item{rangeForAvg}{A vector with the start and ending time period to be used to calculate the
average effective reproductive number.}

\item{B}{The number of bootstrap samples (default is 1000)}

\item{alpha}{The alpha level for the confidence intervals (default is 0.05)}
}
\value{
A list with three dataframes: one with the individual-level reproductive numbers,
 one with the time-level reproductive numbers, and one with the average effective reproductive
 number. ADD MORE DETAILS ON COLUMNS
}
\description{
Uses probabilities to resample the individual-level reproductive numbers in order to
estimate bootstrap confidence intervals for the time-level and average effective
reproductive numbers.
}
\details{
Add details section
}
\examples{
#Insert example here

}
