% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbabilities.R
\name{calcProbabilities}
\alias{calcProbabilities}
\title{Calculates Relative Transmission Probabilities}
\usage{
calcProbabilities(orderedPair, indIDVar, edgeIDVar, goldStdVar, covariates,
  label = "", l = 1, nbWeighting = FALSE, n = 10, m = 1,
  nReps = 10)
}
\arguments{
\item{orderedPair}{The name of the ordered pair-level dataset with the covariates.}

\item{indIDVar}{The variable name (in quotes) of the individual ID variable.}

\item{edgeIDVar}{The variable name (in quotes) of the edge ID variable.}

\item{goldStdVar}{The variable name (in quotes) that will define linking status.}

\item{covariates}{A character vector containing the covariate variable names.}

\item{label}{An optional label string for the run (default is NULL).}

\item{l}{Laplace smoothing parameter that is added to each cell (default is 1).}

\item{nbWeighting}{A logical scalar. Do you want to use deep frequency weighting in NB (default is FALSE)?}

\item{n}{The number of folds for nxm cross valindIDation.}

\item{m}{The number of times to create n folds in nxm cross valindIDation.}

\item{nReps}{The number of times to randomly select one infector.}
}
\value{
List containing two dataframes:
\enumerate{
  \item \code{probabilities} - a dataframe of transmission probabilities with the
   following columns:
     \itemize{
       \item \code{label} - the optional label of the run
       \item \code{pAvg} - the mean transmission probability for the pair over all runs
       \item \code{pSD} - the standard deviation of the transmission probability for the pair
        over all runs
       \item \code{pScaled} - the mean relative transmission probability for the pair over
        all runs: pAvg scaled so that the probabilities for all infectors per infectee add to 1.
       \item \code{pRank} - the rank of the probability of the the pair out of all pairs for that
       infectee
       \item \code{nSamples} - the number of probability estimates that contributed to pAvg. This
       represents the number of validation datasets this pair was included in over the \code{nxm}
       cross validation repeated \code{nReps} times.
       \item \code{edgeIDVar} - the edgeID variable with the name specified
     }
  \item \code{estimates} - a dataframe with the effect estimates with the following columns:
     \itemize{
       \item \code{level} - the covariate name and level
       \item \code{label} - the optional label of the run
       \item \code{ratioMean} - the mean value of the likelihood ratio across runs
       \item \code{ratioMin} - the min value of the likelihood ratio across runs
       \item \code{ratioMax} - the max value of the likelihood ratio across runs
       \item \code{ratioSD} - the standard deviation of the likelihood ratio across runs
       \item \code{nSamples} - the number of samples included in the average: \code{n * m * nReps}
     }
}
}
\description{
\code{calcProbabilities} uses naive Bayes and cross vaindIDation to calculate the relative
 transmission probabilities
}
\details{
Add details section
}
\examples{
#Insert example here

}
