% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcR.R
\name{calcR}
\alias{calcR}
\title{Calculates the effective reproductive number}
\usage{
calcR(probs, indIDVar, dateVar, pVar, timeFrame = c("days", "months",
  "weeks", "years"), rangeForAvg = NULL, bootSamples = 0,
  alpha = 0.05)
}
\arguments{
\item{probs}{The name of the dateset with transmission probabilities}

\item{indIDVar}{The variable name (in quotes) of the individual ID varaibles
(dataframe \code{probs} 
must have variables called \code{<indIDVar>.1} and \code{<indIDVar>.2}).}

\item{dateVar}{The variable name (in quotes) of the dates that the individuals are
observed (dataframe \code{probs}  must have variables called \code{<dateVar>.1} and
\code{<dateVar>.2}).}

\item{pVar}{The variable name (in quotes) of the transmission probabilities.}

\item{timeFrame}{The time frame used to calculate Rt.}

\item{rangeForAvg}{A vector with the start and ending time period to be used to calculate
the average effective reproductive number.}

\item{bootSamples}{The number of bootstrap samples; if 0, then no confidence intervals
are calculated.}

\item{alpha}{The alpha level for the confidence intervals.}
}
\value{
A list with three dataframes: one with the individual-level reproductive numbers,
 one with the time-level reproductive numbers, and one with the average effective reproductive
 number. ADD MORE DETAILS ON COLUMNS
}
\description{
Uses relative transmission probabilities to calculate the individual-level,
time-level and overall average effective reproductive numbers for an outbreak.
}
\details{
Add details section
}
\examples{
#Insert example here

}
