% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcR.R
\name{calcRt}
\alias{calcRt}
\title{Calculates Rt}
\usage{
calcRt(riData, timeFrame = c("days", "weeks", "months", "years"))
}
\arguments{
\item{riData}{The name of the dateset with individual-level reproductive numbers.}

\item{timeFrame}{The time frame used to calculate Rt. One of "days", "weeks", "months", "years".}
}
\value{
Time-level dataframe with the following variables: timeRank = the rank of the time interval
used to calculate the reproductive number and Rt = the reproductive number for that time interval.
}
\description{
Uses probabilities to calculate the time-level effective reproductive number from individual-level
reproductive numbers
}
