% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSI.R
\name{estimateSI}
\alias{estimateSI}
\title{Estimates the serial interval distribution}
\usage{
estimateSI(probs, indIDVar, timeDiffVar, pVar, clustMethod = c("none",
  "n", "kd", "hc_absolute", "hc_relative"), cutoff = NA, initialPars,
  shift = 0, epsilon = 1e-05)
}
\arguments{
\item{probs}{The name of the dateset with transmission probabilities.}

\item{indIDVar}{The name (in quotes) of the individual ID columns
(dataframe \code{probs} must have variables called \code{<indIDVar>.1}
 and \code{<indIDVar>.2}).}

\item{timeDiffVar}{The name (in quotes) of the column with the difference
in time between symptom onset (or its proxy) for the pair of cases. The
units of this variable (hours, days, years) defines the units of the resulting
SI distribution.}

\item{pVar}{The column name (in quotes) of the transmission probabilities.}

\item{clustMethod}{The method used to cluster the infectors
(see \code{\link{clusterInfectors}}).}

\item{cutoff}{The cutoff for clustering (see \code{\link{clusterInfectors}}).}

\item{initialPars}{A vector of length two with the shape and scale 
to initialize the gamma distribution parameters.}

\item{shift}{A value in the same units as \code{timeDiffVar} that the
gamma distribution should be shifted. The default value of 0 is a 
standard gamma distribution.}

\item{epsilon}{The difference between successive estimates of the shape and
scale parameters that indicates convergence.}
}
\value{
A dataframe with one row and the following columns:
\itemize{
   \item \code{nInd} - the number of infectees who have SIs included
   in the SI estimate.
   \item \code{shape} - the shape of the estimated gamma distribution for the SI.
   \item \code{scale} - the scale of the estimated gamma distribution for the SI.
   \item \code{meanSI} - the mean of the estimated gamma distribution for the SI 
   (\code{shape * scale + shift})
   \item \code{medianSI} - the median of the estimated gamma distribution for the SI
   (\code{qgamma(0.5, shape, scale) + shift)})
   \item \code{sdSI} - the standard deviation of the estimated gamma distribution for
   the SI (\code{shape * scale ^ 2})
 }
}
\description{
The function \code{estimateSI} uses the relative transmission probabilities to estimate
the serial interval (SI) distribution assuming a gamma distribution.
It uses the PEM algorithm developed by Hens et al. 2012 extending their method to include
restricting analysis to the top cluster of possible infectors.
}
\details{
The PEM algorithm uses the prior probability that each pair is connected by direct
transmission to estimate the serial interval using estimation maximization. The
current serial interval distribution parameters are used to update the probabilities
which are then used to update the serial interval distribution parameters. The process
continues until the parameters converge (indicated by a change of less than \code{epsilon})
between interations.

This function acts as a wrapper around \code{\link{performPEM}} which integrates
estimation of the SI distribution with clustering the infectors and calculates
derived parameters (mean, median, sd) of the SI distribution. Generally, this function
should be called instead of \code{\link{performPEM}} directly.

All pairs of cases can be used in the estimation process by setting
\code{clustMethod = "none"}. However, if the probabilities are from a algorithm such as
\code{\link{nbProbabilities}}, then it is recommeneded to use a clustering method
and only include the top cluster of infectors for infectees which have such a cluster.
This can be specified by using the \code{clustMethod} and \code{cutoff} arguements which
are passed into \code{\link{clusterInfectors}}. See the details of this function for
a description of the different clustering methods.

The method can be performed with any serial interval distribution,
but this version of this function assumes that the serial interval has a gamma distribution.
The function does allow for a shifted gamma distribution. The \code{shift} arguement
defines how much the gamma distribution should be shifted. Any observed serial intervals
that are less than this shift will have probability 0. This parameter should be used if 
there is a clinically lower bound for the possible serial interval. If this arguement
is not specified then a standard gamma function is used. The units of the
estimated gamma distribution will be defined by the units of the provided
\code{<timeDiffVar>} column. The value of the \code{shift} should be in the same units.

The algorithm requires initial parameters which should be specified as a vector: 
\code{c(<shape>, <scale>)}. These parameters should result in a gamma distribution
that is on the desired scale, set by the \code{<timeDiffVar>} column.
}
\examples{

## Use the pairData dataset which represents a TB-like outbreak
# First create a dataset of ordered pairs
orderedPair <- pairData[pairData$infectionDiffY > 0, ]

## Create a variable called snpClose that will define probable links
# (<3 SNPs) and nonlinks (>12 SNPs) all pairs with between 2-12 SNPs
# will not be used to train.
orderedPair$snpClose <- ifelse(orderedPair$snpDist < 3, TRUE,
                        ifelse(orderedPair$snpDist > 12, FALSE, NA))
table(orderedPair$snpClose)

## Running the algorithm
# NOTE should run with nReps > 1
covariates = c("Z1", "Z2", "Z3", "Z4", "timeCat")
resGen <- nbProbabilities(orderedPair = orderedPair,
                            indIDVar = "individualID",
                            pairIDVar = "pairID",
                            goldStdVar = "snpClose",
                            covariates = covariates,
                            label = "SNPs", l = 1,
                            n = 10, m = 1, nReps = 1)
                            
## Merging the probabilities back with the pair-level data
allProbs <- merge(resGen[[1]], orderedPair, by = "pairID", all = TRUE)

## Estimating the serial interval

# Using all pairs
estimateSI(allProbs, indIDVar = "individualID", timeDiffVar = "infectionDiffY",
pVar = "pScaled", clustMethod = "none", initialPars = c(2, 2))

# Using hierarchical clustering with a 0.05 absolute difference cutoff
estimateSI(allProbs, indIDVar = "individualID", timeDiffVar = "infectionDiffY",
          pVar = "pScaled", clustMethod = "hc_absolute", cutoff = 0.05,
          initialPars = c(2, 2))

# Using a shifted gamma distribution:
# not allowing serial intervals of less than 4 months (0.25 years)
estimateSI(allProbs, indIDVar = "individualID", timeDiffVar = "infectionDiffY",
pVar = "pScaled", clustMethod = "hc_absolute", cutoff = 0.05,
initialPars = c(2, 2), shift = 0.25)


}
\references{
Hens N, Calatayud L, Kurkela S, Tamme T, Wallinga J. Robust reconstruction and
analysis of outbreak data: influenza A (H1N1) v transmission in a school-based 
population. \emph{American Journal of Epidemiology}. 2012 Jul 12;176(3):196-203.
}
\seealso{
\code{\link{nbProbabilities}} \code{\link{clusterInfectors}}
 \code{\link{performPEM}}
}
