% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeResults.R
\name{nbNetwork}
\alias{nbNetwork}
\title{Creates a network of the relative transmission probabilities}
\usage{
nbNetwork(
  df,
  indIDVar,
  dateVar,
  pVar,
  clustMethod = c("none", "n", "kd", "hc_absolute", "hc_relative"),
  cutoff = NA,
  blackAndWhite = FALSE,
  probBreaks = c(-0.01, 0.001, 0.005, 0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 1)
)
}
\arguments{
\item{df}{The name of the dateset with transmission probabilities (column \code{pVar}),
individual IDs (columns \code{<indIDVar>.1} and \code{<indIDVar>.2}), and the dates of
observation (columns \code{<dateVar>.1} and \code{<dateVar>.2}).}

\item{indIDVar}{The name (in quotes) of the individual ID columns
(data frame \code{df} must have variables called \code{<indIDVar>.1}
 and \code{<indIDVar>.2}).}

\item{dateVar}{The name (in quotes) of the columns with the dates that the individuals are
observed (data frame \code{df} must have variables called \code{<dateVar>.1} and
\code{<dateVar>.2}).}

\item{pVar}{The name (in quotes) of the column with transmission probabilities.}

\item{clustMethod}{The method used to cluster the infectors; one of 
\code{"none", "n", "kd", "hc_absolute", "hc_relative"} where \code{"none"} or
not specifying a value means use all pairs with no clustering
(see \code{\link{clusterInfectors}} for detials on clustering methods).}

\item{cutoff}{The cutoff for clustering (see \code{\link{clusterInfectors}}).}

\item{blackAndWhite}{A logical. If \code{TRUE}, then the edges are colored in greyscale,
if \code{FALSE}, then the edges are colored with shades of blue.}

\item{probBreaks}{A numeric vector containing between 3 and 10 elements specifying the
boundaries used to classify the probabilities and color the edges.
The first element should be less than 0 and the last should be 1.}
}
\description{
The function \code{nNetwork} creates a network of the transmission probabilities.
The nodes are the individuals and the edges represent possible transmission pairs.
The darker the edge, the higher the probability that the pair is a transmission link.
If a cluster method is specified using \code{clustMethod} and \code{cutoff}, only edges
that are in the top cluster of infectors will be drawn.
}
\details{
Users have the option of specifying how the probabilities should be grouped into different
color shades through the argument \code{probBreaks}. The probabilities are split into groups by
using \code{probBreaks} as the \code{breaks} argument in \code{\link[base]{cut}} with the default options.
The length of the vector should be between 3 and 10 and the first element should be less than 0 and 
the last 1 so that all probabilities are guarenteed to be classified.
The colors are defined with the code \code{brewer.pal(length(probBreaks) - 1, "Blues")}
(where "Blues" is replaced by "Greys" if \code{blackAndWhite} is set to \code{TRUE}).
}
\examples{

## NOT RUN ##
# ## Network of all pairs in color with the default probability breaks
# par(mar = c(0, 0, 0.2, 0))
# nbNetwork(nbResults, indIDVar = "individualID", dateVar = "infectionDate",
# pVar = "pScaled", clustMethod = "none")

## NOT RUN ##
# ## Network of just the top cluster of infectors, black and white, changing the probability breaks
# par(mar = c(0, 0, 0.2, 0))
# nbNetwork(nbResults, indIDVar = "individualID", dateVar = "infectionDate",
#           pVar = "pScaled", clustMethod = "hc_absolute", cutoff = 0.05,
#           blackAndWhite = TRUE, probBreaks = c(-0.01, 0.01, 0.1, 0.25, 0.5, 1))


}
\seealso{
\code{\link{nbProbabilities}} \code{\link{clusterInfectors}}
}
