% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbabilities.R
\name{calcProbabilities}
\alias{calcProbabilities}
\title{Calculates Relative Transmission Probabilities}
\usage{
calcProbabilities(orderedPair, indIDVar, edgeIDVar, goldStdVar, covariates,
  label = "", l = 1, nbWeighting = FALSE, n = 10, m = 1,
  nReps = 10)
}
\arguments{
\item{orderedPair}{The name of the ordered pair-level dataset with the covariates.}

\item{indIDVar}{The variable name (in quotes) of the individual ID variable.}

\item{edgeIDVar}{The variable name (in quotes) of the edge ID variable.}

\item{goldStdVar}{The variable name (in quotes) that will define linking status.}

\item{covariates}{A character vector containing the covariate variable names.}

\item{label}{An optional label string for the run (default is NULL).}

\item{l}{Laplace smoothing parameter that is added to each cell (default is 1).}

\item{nbWeighting}{A logical scalar. Do you want to use deep frequency weighting in NB (default is FALSE)?}

\item{n}{The number of folds for nxm cross valindIDation.}

\item{m}{The number of times to create n folds in nxm cross valindIDation.}

\item{nReps}{The number of times to randomly select one infector.}
}
\value{
List containing two dataframes: "probs" with pairdata with an extra column with the
   average probabilities over all of the cross validation runs and "coeff" with the
   average, minimum, and maximum coefficient values over all cv runs.
}
\description{
Uses naive Bayes and cross vaindIDation to calculate the relative transmission probabilities
}
\details{
Add details section
}
\examples{
#Insert example here

}
