% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbabilities.R
\name{calcProbabilities}
\alias{calcProbabilities}
\title{Calculates Relative Transmission Probabilities}
\usage{
calcProbabilities(indData, pairData, dateVar, idVar, goldStdVar,
  covariates, label = NULL, n = 10, m = 1, nReps = 50)
}
\arguments{
\item{indData}{individual-level dataset with covariate values}

\item{pairData}{pair-level dataset with covariate values}

\item{dateVar}{the variable name (in quotes) of the date that the individual is observed}

\item{idVar}{the variable name (in quotes) of the id variable}

\item{goldStdVar}{the variable name (in quotes) that will define linking status}

\item{covariates}{vector of the covariate variable names}

\item{label}{optional label for the run (default is NULL)}

\item{n}{number of folds for nxm cross validation}

\item{m}{number of times to create n folds in nxm cross validation}

\item{nReps}{number of times to randomly select one infector}
}
\value{
List containing two dataframes: "probs" with pairdata with an extra column with the
   average probabilities over all of the cross validation runs and "coeff" with the
   average, minimum, and maximum coefficient values over all cv runs.
}
\description{
Uses naive Bayes and cross vaidation to calculate the relative transmission probabilities
}
\examples{
#Insert example here

}
