% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcProbabilities.R
\name{calcProbabilities}
\alias{calcProbabilities}
\title{Calculate Relative Transmission Probabilities}
\usage{
calcProbabilities(indData, pairData, observationDate,
  individualID = "individualID", goldStandard, covariates,
  scheme = NULL, n = 10, m = 1, nReps = 50)
}
\arguments{
\item{indData}{individual-level dataset with covariate values}

\item{pairData}{pair-level dataset with covariate values}

\item{observationDate}{the variable name (in quotes) of the date that the individual is observed}

\item{individualID}{the variable name (in quotes) of the id variable}

\item{goldStandard}{the variable name (in quotes) that will define linking status}

\item{covariates}{vector of the covariate variable names}

\item{scheme}{optional label for the run (default is NULL)}

\item{n}{number of folds for nxm cross validation}

\item{m}{number of times to create n folds in nxm cross validation}

\item{nReps}{number of times to randomly select one infector}
}
\value{
List containing two dataframes: "probs" with pairdata with an extra column with the probabilities and
   "coeff" with the coefficients.
}
\description{
Uses naive Bayes and cross vaidation to calculate the relative transmission probabilities
}
\examples{
#Insert example here

}
