% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcR.R
\name{calcR}
\alias{calcR}
\title{Calculates the effective reproductive number}
\usage{
calcR(probs, dateVar, indIDVar, pVar, timeFrame = c("days", "months",
  "weeks", "years"), rangeForAvg = NULL)
}
\arguments{
\item{probs}{The name of the dateset with transmission probabilities}

\item{dateVar}{The variable name (in quotes) of the dates that the individuals are observed.}

\item{indIDVar}{The variable name (in quotes) of the individual ID.}

\item{pVar}{The variable name (in quotes) of the transmission probabilities.}

\item{timeFrame}{The time frame used to calculate Rt. One of "days", "weeks", "months", "years".}

\item{rangeForAvg}{A vector with the start and ending time period to be used to calculate the
average effective reproductive number.}
}
\value{
A list with three dataframes: one with the individual-level reproductive numbers,
 one with the time-level reproductive numbers, and one with the average effective reproductive
 number. ADD MORE DETAILS ON COLUMNS
}
\description{
Uses probabilities to calculate the individual-level, time-level and overall average
effective reproductive numbers for an outbreak.
}
\details{
Add details section
}
\examples{
#Insert example here

}
