% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcR.R
\name{calcRi}
\alias{calcRi}
\title{Calculates Ri}
\usage{
calcRi(probs, pVar, indIDVar, dateVar)
}
\arguments{
\item{probs}{The name of the dateset with transmission probabilities}
}
\value{
Individual-level dataframe with the following variables: indID, date, Ri,
     iden, nInfectors, nInfectees
}
\description{
Uses probabilities to calculate the individual-level reproductive number
}
