% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRi.R
\name{calcRi}
\alias{calcRi}
\title{Calculates Ri}
\usage{
calcRi(probs, dateVar, idVar, pVar)
}
\arguments{
\item{probs}{dateset with transmission probabilities}

\item{dateVar}{the variable name (in quotes) of the date that the individual is observed}

\item{idVar}{the variable name (in quotes) of the id variable}

\item{pVar}{the variable name (in quotes) of the transmission probabilities}
}
\value{
Individual-level dataframe with the following variables: id, Ri, date, iden, nInfectors,
     month, year, dayR, monthR, yearR
}
\description{
Uses probabilities to calculate the individual level reproductive number
}
\examples{
#Insert example here

}
