% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRi.R
\name{calcRi}
\alias{calcRi}
\title{Calculates Ri}
\usage{
calcRi(probs, dateVar, indIDVar, pVar)
}
\arguments{
\item{probs}{The name of the dateset with transmission probabilities}

\item{dateVar}{The variable name (in quotes) of the date that the individual is observed.}

\item{indIDVar}{The variable name (in quotes) of the individual ID variable.}

\item{pVar}{The variable name (in quotes) of the transmission probabilities.}
}
\value{
Individual-level dataframe with the following variables: indID, Ri, date,
     iden, nInfectors, month, year, dayR, monthR, yearR
}
\description{
Uses probabilities to calculate the individual level reproductive number
}
\details{
Add details section
}
\examples{
#Insert example here

}
