% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcR.R
\name{calcRtAvg}
\alias{calcRtAvg}
\title{Calculates Average Rt}
\usage{
calcRtAvg(rtData, rangeForAvg = NULL)
}
\arguments{
\item{rtData}{The name of the dateset with time-level reproductive numbers.}

\item{rangeForAvg}{A vector with the start and ending time period to be used to calculate the
average effective reproductive number.}
}
\value{
The value of the average time-level reproductive number for the outbreak within
"rangeForAvg"
}
\description{
Uses probabilities to calculate the average time-level effective reproductive number.
}
