% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indToPair.R
\name{indToPair}
\alias{indToPair}
\title{Individual to Pair}
\usage{
indToPair(indData, indIDVar, separator = "_", dateVar = NULL,
  ordered = FALSE)
}
\arguments{
\item{indData}{An individual-level dataframe.}

\item{indIDVar}{The variable name (in quotes) of the individual ID variable.}

\item{separator}{The character to be used to separate the individual IDs when creating
the pairID}

\item{dateVar}{The variable name (in quotes) of the dates that the individuals are observed
(optional unless ordered = TRUE). If supplied, the time difference in days and years between
individuals will be calculated.}

\item{ordered}{A logical indicating if a set of ordered pairs should be returned
(dateVar.1 before dateVar.2). If FALSE a dataframe of all pairs will be returned}
}
\value{
A dataframe of either all possible pairs of individuals (ordered = FALSE) or ordered
pairs of individuals (ordered = TRUE). The dataframe will have all of the original variables
with suffixes ".1" and ".2" corresponding to the original values of 
\code{indIDVar.1} and \code{indIDVar.2}.

Added to the dataframe will be a column called \code{pairID} which is \code{indIDVar.1}
and \code{indIDVar.2} separated by \code{separator}.

If dateVar is provided the dataframe will also include variables \code{timeDiff} and 
\code{timeDiffY} giving the difference of time in days and years respectively between the value
of \code{dateVar} for \code{indIDVar.1} and \code{indIDVar.2}
}
\description{
Transforms a dataset of individuals to a dataset of pairs
}
\details{
Add details section
}
\examples{
#Insert example here

}
