% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performNB.R
\name{performNB}
\alias{performNB}
\title{Perform Naive Bayes}
\usage{
performNB(training, validation, covariates, goldStdVar,
  weighting = FALSE, label)
}
\arguments{
\item{training}{training dataset}

\item{validation}{validation dataset}

\item{covariates}{vector of the covariate variable names}

\item{goldStdVar}{the variable name (in quotes) that will define linking status}

\item{weighting}{should deep frequency covariate weighting be applied}

\item{label}{optional label for the run (default is NULL)}
}
\value{
List containing two dataframes: "probs" with pairdata with an extra column with the
   probabilities and "coeff" with the coefficient values.
}
\description{
Calculates predicted probabilities using naive Bayes
}
\examples{
#Insert example here

}
