% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performNB.R
\name{performNB}
\alias{performNB}
\title{Perform Naive Bayes}
\usage{
performNB(training, validation, edgeIDVar, goldStdVar, covariates, l = 1,
  nbWeighting = FALSE)
}
\arguments{
\item{training}{The training dataset name.}

\item{validation}{The validation dataset name.}

\item{edgeIDVar}{The variable name (in quotes) of the edge ID variable.}

\item{goldStdVar}{The variable name (in quotes) that will define linking status.}

\item{covariates}{A character vector containing the covariate variable names.}

\item{l}{Laplace smoothing parameter that is added to each cell (default is 1).}

\item{nbWeighting}{A logical scalar indicatin if you want to use deep frequency weighting.}
}
\value{
List containing two dataframes: "probs" with pairdata with an extra column with the
   probabilities and "coeff" with the coefficient values.
}
\description{
Calculates predicted probabilities using naive Bayes
}
\details{
Add details section
}
\examples{
#Insert example here

}
