% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performNB.R
\name{performNB}
\alias{performNB}
\title{Perform Naive Bayes}
\usage{
performNB(training, validation, covariates, goldStdVar,
  nbWeighting = FALSE, label)
}
\arguments{
\item{training}{The training dataset name.}

\item{validation}{The validation dataset name.}

\item{covariates}{A character vector containing the covariate variable names.}

\item{goldStdVar}{The variable name (in quotes) that will define linking status.}

\item{nbWeighting}{A logical scalar. Do you want to use deep frequency weighting in NB (default is FALSE)?}

\item{label}{An optional label string for the run (default is NULL).}
}
\value{
List containing two dataframes: "probs" with pairdata with an extra column with the
   probabilities and "coeff" with the coefficient values.
}
\description{
Calculates predicted probabilities using naive Bayes
}
\details{
Add details section
}
\examples{
#Insert example here

}
